backend/requirements.txt
dashscope
fastapi
ffmpeg-python
pydantic
python-dotenv
python-multipart
requests
supabase
uvicorn
certifi
